+++ packages/core/src/upload/upload.test.ts
@@ -198,7 +198,6 @@ describe('UploadService', () => {
         threads: 0,
         sprite: true,
         poster: true,
-        hdrOutput: 'both',
       },
     })
   })
@@ -250,7 +249,6 @@ describe('UploadService', () => {
         threads: 12,
         sprite: true,
         poster: true,
-        hdrOutput: 'both',
       },
     })
   })